// packages/annotator/src/annotator/types.ts
export interface Selector {
	/**
	 * A Selector can be refined by another Selector.
	 *
	 * See {@link https://www.w3.org/TR/2017/REC-annotation-model-20170223/#refinement-of-selection
	 * | §4.2.9 Refinement of Selection} in the Web Annotation Data Model.
	 *
	 * Corresponds to RDF property {@link http://www.w3.org/ns/oa#refinedBy}
	 */
	refinedBy?: Selector;
}

/**
 * The {@link https://www.w3.org/TR/2017/REC-annotation-model-20170223/#css-selector
 * | CssSelector} of the Web Annotation Data Model.
 *
 * Corresponds to RDF class {@link http://www.w3.org/ns/oa#CssSelector}
 *
 * @public
 */
export interface CssSelector extends Selector {
	type: 'CssSelector';
	value: string;
}
// ...
export interface TextPositionSelector extends Selector {
	// more precisely: non-negative integer
	end: number;
	start: number;
	type: 'TextPositionSelector'; // more precisely: non-negative integer
}